Move 'FetchKeys()' function into a subpkg (#433)
Fixes #433
