+++ controller/token.go
@@ -18,6 +18,7 @@ import (
 	"github.com/fabric8-services/fabric8-auth/login"
 	"github.com/fabric8-services/fabric8-auth/rest"
 	"github.com/fabric8-services/fabric8-auth/token"
+	keycloakkeys "github.com/fabric8-services/fabric8-auth/token/keycloak/keys"
 	"github.com/fabric8-services/fabric8-auth/token/link"
 	"github.com/fabric8-services/fabric8-auth/token/provider"
 
@@ -55,11 +56,11 @@ func NewTokenController(service *goa.Service, db application.DB, auth login.Keyc
 
 // Keys returns public keys which should be used to verify tokens
 func (c *TokenController) Keys(ctx *app.KeysTokenContext) error {
-	var publicKeys token.JsonKeys
+	var publicKeys keycloakkeys.JSONKeys
 	if ctx.Format != nil && *ctx.Format == "pem" {
 		publicKeys = c.TokenManager.PemKeys()
 	} else {
-		publicKeys = c.TokenManager.JsonWebKeys()
+		publicKeys = c.TokenManager.JSONWebKeys()
 	}
 
 	return ctx.OK(&app.PublicKeys{Keys: publicKeys.Keys})
